move LiveTimeGraph in widgets.py
