+++ gui/__init__.py
@@ -1,1 +1,1 @@
-from gui.widgets import CommandButtons, MessageBox, GatewayStatus
+from gui.widgets import CommandButtons, MessageBox, GatewayStatus, LiveTimeGraph
